feat(choice): add segment component
This PR adds the segment component which has the same functionality
as the radio button. For this reason the radio group and button have
been made generic and renamed to 'choice'.

BREAKING CHANGE: all the references to lg-radio-group and lg-radio-component
should be changed to lg-choice-group and lg-choice-component.
The default variant is 'radio'.